\name{ODMconverter-package}
\alias{ODMconverter-package}
\docType{package}
\title{
ODMconverter
}
\description{
Transformation of ODM files into R format, office format, CDA format and vice versa.\cr
CDISC ODM is very commonly used in clinical trials.\cr
Semantic annotation (such as UMLS, SNOMED or LOINC codes) are transformed accordingly by those conversion tools.
}
\author{
Martin Dugas\cr
Maintainer: Martin Dugas <dugas@uni-muenster.de>
}
\references{
see functions ODM2office, office2ODM, ODM2R, R2ODM, ODM2CDA, CDA2ODM
}
\keyword{ ODM }
\keyword{ CDA }